+++ components/static/Detail.vue
@@ -5,7 +5,7 @@
     </el-empty>
   </div>
   <div v-else class="app-container">
-    <aside v-if="treeData && treeData.length > 0">
+    <aside v-if="formData.setting.docTreeEnabled && treeData && treeData.length > 0">
       <!-- 固定的图标 -->
       <div class="sidebar-toggle" @click="toggleSidebar">
         <el-icon v-if="isSidebarVisible">
@@ -56,7 +56,7 @@
         </div>
       </div>
     </main>
-    <aside v-if="outlineData && outlineData.length > 0" class="floating-toc">
+    <aside v-if="formData.setting.outlineEnabled && outlineData && outlineData.length > 0" class="floating-toc">
       <outline :outline-data="outlineData" :max-depth="outlineMaxDepth" />
     </aside>
   </div>
@@ -72,7 +72,8 @@ import { useAuthModeFetch } from "~/composables/useAuthModeFetch"
 import { useProviderMode } from "~/composables/useProviderMode"
 import Sidebar from "~/components/static/Sidebar.vue"
 import Outline from "~/components/static/Outline.vue"
-import { Fold, Expand } from "@element-plus/icons-vue"
+import { Expand, Fold } from "@element-plus/icons-vue"
+import { useStaticSettingStore } from "~/stores/useStaticSettingStore"
 
 // https://github.com/nuxt/nuxt/issues/15346
 // 由于布局是个宏，静态构建情况下，不能动态设置，只能在前面的页面写死
@@ -94,10 +95,12 @@ const id = props.pageId ?? ((route.params.id ?? "") as string)
 const { getFirstImageSrc } = useServerAssets()
 const { fetchPostMeta } = useAuthModeFetch()
 const { providerMode } = useProviderMode()
+const { getStaticSetting } = useStaticSettingStore()
 
 // datas
 const formData = reactive({
   post: {} as Post,
+  setting: {} as any,
   shareEnabled: true,
   isExpires: false,
 })
@@ -114,7 +117,15 @@ const getPostData = async () => {
   const attrs = JsonUtil.safeParse<any>(formData.post?.attrs ?? "{}", {})
   formData.isExpires = checkExpires(attrs)
 }
+const getSetting = async () => {
+  const currentSetting = await getStaticSetting()
+  logger.info("currentSetting=>", currentSetting)
+  // 默认没有设置的时候应该显示
+  formData.setting.docTreeEnabled = currentSetting?.docTreeEnabled ?? true
+  formData.setting.outlineEnabled = currentSetting?.outlineEnabled ?? true
+}
 await getPostData()
+await getSetting()
 
 if (!props.overrideSeo) {
   const titleSign = " - " + t("blog.share")
@@ -226,7 +237,7 @@ onMounted(() => {})
   position fixed
   top 20px
   right 20px
-  min-idth 200px
+  min-width 200px
   max-width 350px
   background-color #fff
   border 1px solid #ddd