+++ composables/useAuthModeFetch.ts
@@ -25,13 +25,15 @@
 
 import { createAppLogger } from "~/common/appLogger"
 import { useSiyuanApi } from "~/composables/api/useSiyuanApi"
+import { JsonUtil, StrUtil } from "zhi-common"
+import { isDev } from "~/common/Constants"
 
 export const useAuthModeFetch = () => {
   const logger = createAppLogger("use-auth-mode-fetch")
   const { kernelApi } = useSiyuanApi()
   const env = useRuntimeConfig()
   const route = useRoute()
-
+  const isSSR = process.server
   /**
    * 获取文本
    *
@@ -99,10 +101,10 @@ export const useAuthModeFetch = () => {
   /**
    * 远程获取配置文本
    *
+   * @param docId - 文档ID
    * @param filename - 文件名
    */
-  const fetchProviderConfigForCurrentUser = async (filename: string): Promise<string> => {
-    const id = (route.params.id ?? "") as string
+  const fetchProviderConfigForCurrentUser = async (docId: string, filename: string): Promise<string> => {
     const apiBase = env.public.providerUrl
     const url = `/api/settings/share`
     const reqUrl = `${apiBase}${url}`
@@ -115,7 +117,7 @@ export const useAuthModeFetch = () => {
       },
       body: JSON.stringify({
         group: "GENERAL",
-        docId: id,
+        docId: docId,
         key: filename,
       }),
     })
@@ -123,6 +125,54 @@ export const useAuthModeFetch = () => {
     logger.info("fetch config text in provider mode finish=>", { resText: resText })
     if (!res.ok) {
       throw new Error("fetch provider config error")
+    }
+    return resText
+  }
+
+  const fetchProviderConfigByAuthorForCurrentUser = async (author: string, filename: string): Promise<string> => {
+    const apiBase = env.public.providerUrl
+    const url = `/api/settings/byAuthor`
+    const reqUrl = `${apiBase}${url}`
+    let resText = ""
+    logger.info(`fetch config text ${filename} in provider mode, reqUrl=>${reqUrl}`)
+    const res = await fetch(reqUrl, {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify({
+        group: "GENERAL",
+        author: author,
+        key: filename,
+      }),
+    })
+    resText = await res.text()
+    logger.info("fetch config text in provider mode finish=>", { resText: resText })
+    if (!res.ok) {
+      throw new Error("fetch provider config error")
+    }
+    return resText
+  }
+
+  const fetchProviderConfigByResource = async (filename: string): Promise<string> => {
+    const apiBase = env.public.providerUrl
+    const url = `/api/settings/byResource`
+    const reqUrl = `${apiBase}${url}`
+    let resText = ""
+    logger.info(`fetch resource ${filename} in provider mode, reqUrl=>${reqUrl}`)
+    const res = await fetch(reqUrl, {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify({
+        key: filename,
+      }),
+    })
+    resText = await res.text()
+    logger.info("fetch resource in provider mode finish=>", { resText: resText })
+    if (!res.ok) {
+      throw new Error("fetch resource error")
     }
     return resText
   }
@@ -140,13 +190,43 @@ export const useAuthModeFetch = () => {
     return resText
   }
 
+  const getAuthorByDomainWhiteList = async (): Promise<string> => {
+    // 先查找 domain 白名单
+    const domainsFile = isDev ? "domains.local.json" : "domains.json"
+    const domainsText = await fetchProviderConfigByResource(domainsFile)
+    const domainsJson = JsonUtil.safeParse<any>(domainsText, {})
+    const domains = domainsJson.domains ?? []
+    // 获取当前页面的 origin
+    // https://stackoverflow.com/a/77175631/4037224
+    const requestURL = useRequestURL()
+    const currentOrigin = StrUtil.isEmptyString(requestURL.origin) ? window.location.origin : requestURL.origin
+    logger.info("current origin=>", currentOrigin)
+    // 查找匹配的 domain 并获取 author
+    const matchedDomain = domains.find((domain: any) => domain.domain === currentOrigin)
+    return matchedDomain ? matchedDomain.author : null
+  }
+
   const fetchConfig = async (filename: string, providerMode: boolean): Promise<string> => {
     console.log("providerMode=>", providerMode)
     let resText: string
     if (providerMode) {
       logger.info(`fetch config text ${filename} in provider mode`)
       try {
-        resText = await fetchProviderConfigForCurrentUser(filename)
+        const docId = (route.params.id ?? "") as string
+        if (docId == "") {
+          // 首页
+          const whiteListAuthor = await getAuthorByDomainWhiteList()
+          if (whiteListAuthor) {
+            logger.info("use author from domain white list for home page")
+            resText = await fetchProviderConfigByAuthorForCurrentUser(whiteListAuthor, filename)
+          } else {
+            logger.info("use default author for home page")
+            resText = await fetchProviderConfigForCurrentUser(docId, filename)
+          }
+        } else {
+          resText = await fetchProviderConfigForCurrentUser(docId, filename)
+        }
+
         logger.info("success fetch config in  provider mode")
       } catch (e) {
         logger.warn("cannot find setting for current user, use default")
@@ -157,6 +237,10 @@ export const useAuthModeFetch = () => {
       resText = await fetchPublicText(filename)
     }
     logger.info("finally resText by fetchConfig=>", { resText: resText })
+    // 存一份到客户端使用
+    if (!isSSR) {
+      window.localStorage.setItem(filename, resText)
+    }
     return resText
   }